move tests into component specific helper
